WIP: Verifying that the teams are migrated correctly

+++ db/schema.rb
@@ -361,11 +361,11 @@ ActiveRecord::Schema.define(:version => 20141015182230) do
     t.datetime "updated_at",                                     :null => false
     t.string   "website"
     t.text     "about"
-    t.integer  "total",                    :default => 0
+    t.float    "total",                    :default => 0.0
     t.integer  "size",                     :default => 0
-    t.integer  "mean",                     :default => 0
-    t.integer  "median",                   :default => 0
-    t.integer  "score",                    :default => 0
+    t.float    "mean",                     :default => 0.0
+    t.float    "median",                   :default => 0.0
+    t.float    "score",                    :default => 0.0
     t.string   "twitter"
     t.string   "facebook"
     t.string   "slug"
